WEB: Resolve static analysis errors
